requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Data Processing
pandas==2.1.3
numpy==1.24.3
sqlalchemy==2.0.23

# AI/ML Libraries
openai==1.3.7
langchain==0.0.350
transformers==4.36.0
torch==2.1.1
scikit-learn==1.3.2
spacy==3.7.2

# Database
sqlite3
psycopg2-binary==2.9.9

# Caching and Session Management
redis==5.0.1
python-multipart==0.0.6

# HTTP and WebSocket
httpx==0.25.2
websockets==12.0

# Utilities
python-dotenv==1.0.0
pydantic-settings==2.1.0
loguru==0.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0
